+++ module/account/src/Domain/Mail/UserPasswordChangedMail.php
@@ -6,31 +6,28 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\Account\Domain;
+namespace Ergonode\Account\Domain\Mail;
 
 use Ergonode\Account\Domain\Entity\User;
-use Ergonode\Account\Domain\ValueObject\ResetToken;
 use Ergonode\Mailer\Domain\Mail;
 use Ergonode\Mailer\Domain\Recipient;
 use Ergonode\Mailer\Domain\Sender;
 use Ergonode\Mailer\Domain\Template;
 use Ergonode\SharedKernel\Domain\Collection\EmailCollection;
 
-class ResetTokenMail extends Mail
+class UserPasswordChangedMail extends Mail
 {
-    public function __construct(User $user, ResetToken $token, string $url)
+    public function __construct(User $user)
     {
         $recipient = new Recipient(new EmailCollection([$user->getEmail()]));
         $template = new Template(
-            '@ErgonodeAccount/message/token.html.twig',
+            '@ErgonodeAccount/message/password_changed.html.twig',
             $user->getLanguage(),
             [
-                'url' => $url,
-                'token' => $token->getValue(),
                 'firstName' => $user->getFirstName(),
             ]
         );
 
-        parent::__construct($recipient, new Sender(), $template, 'Reset your password at Ergonode PIM');
+        parent::__construct($recipient, new Sender(), $template, 'You’ve changed your password at Ergonode PIM!');
     }
 }
