+++ module/account/src/Resources/translations/mailer.en.yaml
@@ -5,3 +5,4 @@ and follow the instructions to reset your password.: and follow the instructions
 If it’s not you who asked for password reset, please ignore this email.: If it’s not you who asked for password reset, please ignore this email.
 Thanks!: Thanks!
 Team Ergonode: Team Ergonode
+You successfully changed your password to your account at Ergonode PIM.: You successfully changed your password to your account at Ergonode PIM.
