+++ module/account/src/Resources/translations/mailer.pl.yaml
@@ -5,3 +5,4 @@ and follow the instructions to reset your password.: i postępować zgodnie z in
 If it’s not you who asked for password reset, please ignore this email.: Jeśli nie Ty prosiłeś o reset hasła, zignoruj tego maila.
 Thanks!: Dziękujemy!
 Team Ergonode: Zespół Ergonode
+You successfully changed your password to your account at Ergonode PIM.: Pomyślnie zmieniono hasło dla Twojego konta w Ergonode PIM.
