+++ module/account/src/Resources/views/message/password_changed.html.twig
@@ -1,0 +1,38 @@
+{% trans_default_domain 'mailer' %}
+
+{% extends "@ErgonodeMailer/layout/default.html.twig" %}
+
+{% block content %}
+    <table width="600px" border="0" cellspacing="0" cellpadding="0" align="center" bgcolor="#fff"
+           style="background-color=#fff; margin-top:40px; margin-bottom:40px; padding:40px;">
+        <tr style="background-color:#fff; color:#33373E;">
+            <td align="center" style="text-align: justify;">
+                <p style="margin-top:16px;">
+                    {{ 'Hey'|trans }} {{ firstName }},
+                </p>
+            </td>
+        </tr>
+
+        <tr style="background-color:#fff; color:#33373E;">
+            <td align="center" style="text-align: justify;">
+                <p style="margin-top:16px;">
+                    {{ 'You successfully changed your password to your account at Ergonode PIM.' |trans }}
+                </p>
+            </td>
+        </tr>
+        <tr style="background-color:#fff; color:#33373E;">
+            <td align="center" style="text-align: justify;">
+                <p style="margin-top:16px;">
+                    {{ 'Thanks!'|trans  }}
+                </p>
+            </td>
+        </tr>
+        <tr style="background-color:#fff; color:#33373E;">
+            <td align="center" style="text-align: justify;">
+                <p style="margin-top:16px;">
+                    {{ 'Team Ergonode'|trans }}
+                </p>
+            </td>
+        </tr>
+    </table>
+{% endblock %}
